requirements.txt
pandas
openpyxl
PyYAML
python-docx
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pytest
